Refactor account component names for consistency
